Update filenames. Re-organzie function localtions.
