+++ projection/validator_delegation/delegation.go
@@ -4,9 +4,127 @@ import (
 	"fmt"
 
 	"github.com/crypto-com/chain-indexing/appinterface/rdb"
+	"github.com/crypto-com/chain-indexing/external/utctime"
 	"github.com/crypto-com/chain-indexing/projection/validator_delegation/utils"
+	"github.com/crypto-com/chain-indexing/projection/validator_delegation/view"
 	"github.com/crypto-com/chain-indexing/usecase/coin"
 )
+
+// handleDelegate - Handle create/update delegation. Return newShares and error.
+func (projection *ValidatorDelegation) handleDelegate(
+	rdbTxHandle *rdb.Handle,
+	height int64,
+	validatorAddress string,
+	delegatorAddress string,
+	amount coin.Int,
+) (coin.Dec, error) {
+
+	validatorsView := NewValidators(rdbTxHandle)
+	delegationsView := NewDelegations(rdbTxHandle)
+
+	// Get the validator
+	validator, found, err := validatorsView.FindByOperatorAddr(validatorAddress, height)
+	if err != nil {
+		return coin.ZeroDec(), fmt.Errorf("error validatorsView.FindByOperatorAddr(): %v", err)
+	}
+	if !found {
+		return coin.ZeroDec(), fmt.Errorf("error validator not found: %v, %v", validatorAddress, height)
+	}
+
+	// Create/Get Delegation
+	delegation, found, err := delegationsView.FindBy(
+		validatorAddress,
+		delegatorAddress,
+		height,
+	)
+	if err != nil {
+		return coin.ZeroDec(), fmt.Errorf("error delegationsView.FindBy(): %v", err)
+	}
+	if !found {
+		delegation = view.DelegationRow{
+			Height:           height,
+			ValidatorAddress: validatorAddress,
+			DelegatorAddress: delegatorAddress,
+			Shares:           coin.ZeroDec(),
+		}
+
+		// Insert an DelegationRow record
+		if errAddDelRecord := delegationsView.Insert(delegation); errAddDelRecord != nil {
+			return coin.ZeroDec(), fmt.Errorf("error delegationsView.Insert(): %v", errAddDelRecord)
+		}
+	}
+
+	// Calculate new shares
+	newShares, err := utils.ValidatorIssueNewShares(validator, amount)
+	if err != nil {
+		return coin.ZeroDec(), fmt.Errorf("error IssueNewShares(): %v", err)
+	}
+
+	// Update Delegation
+	delegation.Shares = delegation.Shares.Add(newShares)
+	if err := delegationsView.Update(delegation); err != nil {
+		return coin.ZeroDec(), fmt.Errorf("error delegationsView.Update(): %v", err)
+	}
+
+	// Update Validator
+	validator.Tokens = validator.Tokens.Add(amount)
+	validator.Shares = validator.Shares.Add(newShares)
+	if err := validatorsView.Update(validator); err != nil {
+		return coin.ZeroDec(), fmt.Errorf("error validatorsView.Update(): %v", err)
+	}
+
+	return newShares, nil
+}
+
+func (projection *ValidatorDelegation) handleUndelegate(
+	rdbTxHandle *rdb.Handle,
+	height int64,
+	validatorAddress string,
+	delegatorAddress string,
+	amount coin.Int,
+	completionTime utctime.UTCTime,
+) error {
+
+	shares, err := projection.unbondAmountToShares(
+		rdbTxHandle,
+		height,
+		validatorAddress,
+		delegatorAddress,
+		amount,
+	)
+	if err != nil {
+		return fmt.Errorf("error projection.unbondAmountToShares(): %v", err)
+	}
+
+	returnAmount, err := projection.unbond(
+		rdbTxHandle,
+		height,
+		validatorAddress,
+		delegatorAddress,
+		shares,
+	)
+	if err != nil {
+		return fmt.Errorf("error projection.unbond(): %v", err)
+	}
+
+	ubd, err := projection.setUnbondingDelegationEntry(
+		rdbTxHandle,
+		delegatorAddress,
+		validatorAddress,
+		height,
+		completionTime,
+		returnAmount,
+	)
+	if err != nil {
+		return fmt.Errorf("error projection.setUnbondingDelegationEntry(): %v", err)
+	}
+
+	if err := projection.insertUBDQueue(rdbTxHandle, ubd, completionTime); err != nil {
+		return fmt.Errorf("error projection.insertUBDQueue(): %v", err)
+	}
+
+	return nil
+}
 
 func (projection *ValidatorDelegation) unbondAmountToShares(
 	rdbTxHandle *rdb.Handle,
